Semester 2/Data Engineering/Project 1/requirements.txt
python-telegram-bot==22.1
requests==2.32.3
python-dotenv==1.1.0